requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bidict==0.23.1
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
motor==3.5.1
multidict==6.0.5
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-engineio==4.9.1
python-jose==3.3.0
python-socketio==5.11.3
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
wsproto==1.2.0
yarl==1.9.4
